+++ backend/src/routes/event-products.routes.ts
@@ -5,6 +5,7 @@ import {
     deleteEventProduct,
     getEventProductById,
     listEventProducts,
+    reorderEventProducts,
     updateEventProduct
 } from '../controllers/event-products.controller';
 import { authMiddleware } from '../middlewares/auth.middleware';
@@ -18,5 +19,6 @@ eventProductsRouter.get('/:epId', asyncHandler(getEventProductById));
 eventProductsRouter.use(asyncHandler(authMiddleware));
 
 eventProductsRouter.post('/', asyncHandler(createEventProduct));
+eventProductsRouter.patch('/reorder', asyncHandler(reorderEventProducts));
 eventProductsRouter.patch('/:epId', asyncHandler(updateEventProduct));
 eventProductsRouter.delete('/:epId', asyncHandler(deleteEventProduct));